Renamed views_logic to mixins
